Load in statehook
